feat: 배터리 정보 얻기

+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/model/LocationInfo.kt
@@ -14,7 +14,7 @@ data class LocationInfo(
     val gyrosensor: List<Float>,
     val directionsensor: List<Float>,
     val lightsensor: List<Float>,
-    val battery: Float,
+    val battery: Int,
     val isInternetOn: Boolean,
     val isGpsOn: Boolean,
     val isRingstoneOn: Boolean
